backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.13.1",
    "anyio==4.4.0",
    "bcrypt==4.0.1",
    "emails==0.6",
    "fastapi==0.111.0",
    "httpx==0.27.0",
    "jinja2==3.1.4",
    "passlib[bcrypt]==1.7.4",
    "psycopg2-binary==2.9.9",
    "pydantic==2.7.4",
    "pydantic-settings==2.3.4",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.9",
    "reportlab==4.2.0",
    "sqlalchemy==2.0.30",
    "sqlmodel==0.0.16",
    "uvicorn[standard]==0.30.1",
]
